# openapi.yaml
openapi: 3.0.3
info:
  title: VibeCheck - Aesthetic Social Media Companion API
  description: |
    API specification for VibeCheck, an aesthetic social media companion application.
    Users can curate content across multiple categories (cinema, music, games, books, travel),
    share their aesthetic identity (aura), earn badges through gamification, and discover
    like-minded users in aesthetic communities.
  version: 1.0.0
  contact:
    name: VibeCheck Support
  license:
    name: MIT
servers:
  - url: https://api.vibecheck.com/v1
    description: Production server
  - url: https://staging-api.vibecheck.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
  /auth/login:
    post:
      tags:
        - User Profile
      summary: User login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
            example:
              email: user@example.com
              password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                user:
                  userId: "u_123"
                  email: user@example.com
                  username: aesthetic_anna
                  role: normal
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Email not verified. Please verify your email before logging in.
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/moderator-login/{token}:
    get:
      tags:
        - User Profile
      summary: Moderator magic-link login
      description: |
        Exchanges a moderator magic-link token for a normal JWT session.
        The token is single-use and expires after 1 hour.
      operationId: moderatorLogin
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Moderator magic-link token from /mod/{token}
      responses:
        '200':
          description: Moderator login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                user:
                  userId: "u_mod123"
                  email: moderator@example.com
                  username: mod_anna
                  role: moderator
        '401':
          description: Invalid or expired moderator link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/register:
    post:
      tags:
        - User Profile
      summary: User registration
      description: User registration requires email verification. A verification email will be sent to the registered email address.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - username
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
      responses:
        '201':
          description: Registration successful. Verification email sent. User must verify email before being able to login.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  emailVerificationRequired:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email or username already exists
  /auth/verify-email:
    get:
      tags:
        - User Profile
      summary: Verify user email
      description: Verify email address using the verification token from the email link. User clicks the link in the email which contains the token as a URL parameter.
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Email verification token received in the verification email
      responses:
        '200':
          description: Email verified successfully. Shows verification confirmation page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                    description: JWT authentication token
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid or expired verification token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/forgot-password:
    post:
      tags:
        - User Profile
      summary: Request password reset
      description: Send password reset email with verification token to user's registered email
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: Email address associated with the account
            example:
              email: user@example.com
      responses:
        '200':
          description: Password reset email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User with this email not found
  /auth/reset-password:
    post:
      tags:
        - User Profile
      summary: Reset user password
      description: Reset password using the reset token sent via email. Email verification is mandatory for password reset.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                  description: Password reset token received in the password reset email
                newPassword:
                  type: string
                  format: password
                  description: New password for the account
            example:
              token: "reset_token_abc123xyz"
              newPassword: "newPassword123"
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    User:
      type: object
      required:
        - userId
        - email
        - username
        - role
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [normal, moderator]
          description: Authorization role of the user account
        avatar:
          type: string
          format: uri
        bio:
          type: string
        emailVerified:
          type: boolean
          description: Whether the user's email has been verified
        socialMediaLinks:
          type: array
          description: Array of user's social media links
          items:
            type: object
            properties:
              platform:
                type: string
                enum: [instagram, twitter, tiktok, youtube, facebook, linkedin, pinterest, spotify, twitch, other]
              url:
                type: string
                format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT